refactor: move tests related to denials
